backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
selenium==4.22.0
webdriver-manager==4.0.1
pandas==2.2.2
openpyxl==3.1.5
